beginnings of a DOM structure.
puttings the rendering code in one main function for now. When it works
and seems to reasonably validate the design the core module I will break
it into another file and leave the core as the simple shell that is
should be. At this time it will be necessary to work on the build stuff.

So some sort of design is emerging.