at least i'm rendering something.
wip: row padding config. So begin the unexpected consequences. If the
user is restrained to being to render a row at a time, then there has to
be config relating to the relationship between rows.

I guess the next step is to try pulling out the rendering code into
another file. Proof of concept? Done.