wip, refactoring and more clearly defining the API.
A few things:

1. Going to try to avoid holding on to a copy of the data within a
closure. There is a function for getting the data bound to the DOM. This
way there will be nothing to keep and sync and should simply things when
it comes to sorting and whatnot.

2. The API is going to be either oncoprint(config, div) -> API functions
such as API row, toggle_whitespace, etc. And the "curried version,"
oncoprint(config) -> (div -> API functions). Not a huge difference, but
it does allow the user to configure an OncoPrint and then re-use it on a
number of divs.

3. It's good that rendering functions need a configuration. This is how
interaction is going to work. When a user changes a setting, you will
get a slightly different rendering function for the current state.

4. The previous point does not account for the transitions that will
have to take place. So when there is a user interaction, I will have to
manually transition the visualization around, as well as update the
configuration for future rendering actions. I don't think having two
steps is that bad. This is basically saying, "No, stop, don't re-render
everything, just do the following changes, and keep track of the
configuration for the future."

5. I'm not sure yet where this transition actions should go because they
will require having a knowledge of the DOM structure of what has been
rendered.